Split some files apart
