+++ src/git-cmd-api.js
@@ -1,0 +1,37 @@
+export async function getHeadForRepo(targetDirname) {
+
+}
+
+export async function getOriginForRepo(targetDirname) {
+}
+
+export async function checkoutSha(targetDirname, sha) {
+}
+
+export function updateRefspecToPullPRs(repository) {
+}
+
+export async function cloneRepo(url, targetDirname, token=null, bare=true) {
+
+}
+
+export async function fetchRepo(targetDirname, token=null, bare=true) {
+
+}
+
+export async function cloneOrFetchRepo(url, checkoutDir, token=null) {
+
+}
+
+export async function resetOriginUrl(target, url) {
+
+}
+
+export async function addFilesToGist(repoUrl, targetDir, artifactDir, token=null) {
+
+}
+
+export async function pushGistRepoToMaster(targetDir, token) {
+
+}
+
